Rename Shopper Insights package name
