+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/EligiblePaymentMethodDetails.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 import org.json.JSONObject
