+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/EligiblePaymentsCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 import java.lang.Exception
