+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/ShopperInsightsInfo.kt
@@ -1,4 +1,6 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
+
+import com.braintreepayments.api.ExperimentalBetaApi
 
 /**
  * Data class encapsulating the result of a shopper insight api request.
