+++ ShopperInsights/src/test/java/com/braintreepayments/api/shopperinsights/EligiblePaymentsApiResultUnitTest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 import org.junit.Assert.*
 import org.junit.Test
