+++ ShopperInsights/src/test/java/com/braintreepayments/api/shopperinsights/EligiblePaymentsApiUnitTest.kt
@@ -1,5 +1,6 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
+import com.braintreepayments.api.ExperimentalBetaApi
 import com.braintreepayments.api.core.BraintreeClient
 import com.braintreepayments.api.core.Configuration
 import com.braintreepayments.api.core.ConfigurationCallback
